[BE-2] Fixed unusable coordinate segment bleeding
